refactor fetchSourceGithub into smaller components
