+++ app-web/plugins/gatsby-source-github-all/utils/sources/github/githubFileTransformer.js
@@ -1,0 +1,43 @@
+/*
+Copyright 2018 Province of British Columbia
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at 
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+Created by Patrick Simonian
+*/
+const { PERSONAS_LIST } = require('../../constants');
+const { fileTransformer } = require('../../transformer');
+const {
+  markdownFrontmatterPlugin,
+  pagePathPlugin,
+  markdownUnfurlPlugin,
+  markdownResourceTypePlugin,
+  externalLinkUnfurlPlugin,
+  markdownPersonaPlugin,
+  repositoryResourcePathPlugin,
+} = require('../../plugins');
+
+const transfomer = (extension, file) => {
+  const ft = fileTransformer(extension, file);
+
+  return ft
+    .use(markdownFrontmatterPlugin)
+    .use(pagePathPlugin)
+    .use(markdownUnfurlPlugin)
+    .use(markdownResourceTypePlugin)
+    .use(externalLinkUnfurlPlugin)
+    .use(markdownPersonaPlugin, { personas: PERSONAS_LIST })
+    .use(repositoryResourcePathPlugin);
+};
+
+module.exports = transfomer;
